+++ Classes/IOCAccountsController.m
@@ -77,7 +77,7 @@
 	self.accountsByEndpoint = [NSMutableDictionary dictionary];
 	for (GHAccount *account in self.accounts) {
 		NSString *endpoint = account.endpoint;
-		if (!endpoint || endpoint.isEmpty) endpoint = kGitHubEndpointURL;
+		if (!endpoint || endpoint.isEmpty) endpoint = kGitHubComURL;
 		if (!self.accountsByEndpoint[endpoint]) {
 			self.accountsByEndpoint[endpoint] = [NSMutableArray array];
 		}